filename creation centralized
optional backup creation with
--no-apk
--no-data
--no-ext-data
--no-keystore
--no-perms

+++ backup_apps.sh
@@ -12,15 +12,43 @@ SYSTEM_PATTERN=""
 SINGLE_APP=""
 DO_ONLY_MATCHING_APPS=false
 MATCHING_APPS=""
-DO_BACKUP_EXTRA_DATA=false
+DO_IT=true # for try run
+DO_ACTION_APK=true
+DO_ACTION_DATA=true
+DO_ACTION_EXT_DATA=true
+DO_ACTION_KEYSTORE=true
+DO_ACTION_PERMISSIONS=true
+HAS_CUSTOM_BACKUP_DIR=false
 
 argCount=${#@}
 while [ $argCount -gt 0 ] ; do
-    if [[ "$1" == "--local" ]]; then
+    if [[ "$1" == "--backup-dir" ]]; then
+        HAS_CUSTOM_BACKUP_DIR=true
+        shift; let argCount-=1
+        BACKUP_DIR="$1"
+        shift; let argCount-=1
+    elif [[ "$1" == "--local" ]]; then
         IS_LOCAL=true
         CUSTOM_BUSYBOX_TARGET_BIN=/tmp/busybox
         shift; let argCount-=1
-
+    elif [[ "$1" == "--do-nothing" ]]; then
+        shift; let argCount-=1
+        DO_IT=false
+    elif [[ "$1" == "--no-apk" ]]; then
+        shift; let argCount-=1
+        DO_ACTION_APK=false
+    elif [[ "$1" == "--no-data" ]]; then
+        shift; let argCount-=1
+        DO_ACTION_DATA=false
+    elif [[ "$1" == "--no-ext-data" ]]; then
+        shift; let argCount-=1
+        DO_ACTION_EXT_DATA=false
+    elif [[ "$1" == "--no-keystore" ]]; then
+        shift; let argCount-=1
+        DO_ACTION_KEYSTORE=false
+    elif [[ "$1" == "--no-perms" ]]; then
+        shift; let argCount-=1
+        DO_ACTION_PERMISSIONS=false
     elif [[ "$1" == "--data-path" ]]; then
         IS_LOCAL=true
         #CUSTOM_BUSYBOX_TARGET_BIN=/tmp/busybox
@@ -88,10 +116,12 @@ updateBusybox
 
 #pushBusybox
 
-#mkBackupDir
-DIR=test
-mkdir -p $DIR
-pushd $DIR
+if ! $HAS_CUSTOM_BACKUP_DIR ; then
+    einfo2 mkBackupDir
+    #mkBackupDir
+fi
+mkdir -p $BACKUP_DIR
+pushd $BACKUP_DIR
 
 if $IS_LOCAL ; then
     PACKAGES=$(xmlstarlet select -T -t -m "//packages/package"  -v "@codePath" -o "|" -v "@name" -n $DATA_PATH/system/packages.xml)
@@ -107,10 +137,22 @@ echo "## Pull apps"
 function matchApp()
 {
     if $DO_ONLY_MATCHING_APPS ; then
-        echo $PACKAGES | tr " " "\n" | egrep "($MATCHING_APPS)$"
+        NEW_APPS=()
+        for x in `echo $MATCHING_APPS | sed -e 's@|@ @g'` ; do
+            NEW_APPS+=(`echo "${PACKAGES}" | grep "${x}$"`)
+        done
+        #echo $PACKAGES | tr " " "\n" | egrep "($MATCHING_APPS)$"
+        PACKAGES="${NEW_APPS[*]}"
+        echo $PACKAGES
     else
         echo $PACKAGES | tr " " "\n" | grep "${PATTERN}" | grep "$SINGLE_APP"
     fi
+}
+
+function getPermsXmlData()
+{
+    local appSign="$1"
+    sudo xmlstarlet sel -t -c "` echo "/runtime-permissions/pkg[@name = 'FOLLER']" | sed -e "s@FOLLER@$appSign@g"`" $DATA_PATH/system/users/0/runtime-permissions.xml
 }
 
 DATADIR=""
@@ -135,29 +177,41 @@ for APP in `matchApp`; do
 	echo $dataDir
 
         if $IS_LOCAL ; then
+                mkdir "${dataDir}" # dir per app
+                appPackage="$(getAppFileName "${dataDir}")"
                 # get apk
-		sudo $BUSYBOX tar -C $DATA_PATH/$appDir -czpf - ./ 2>/dev/null | pv -trabi 1 > app_${dataDir}.tar.gz
-                # get data
-		sudo $BUSYBOX tar -C $DATA_PATH/data/$dataDir -czpf - ./ 2>/dev/null | pv -trabi 1 > data_${dataDir}.tar.gz
-
-                # get keystore if exists
-                AS=""
-                USERID="`getUserId  $DATA_PATH/data/$dataDir`"
-                keystorePath=$DATA_PATH/misc/keystore/user_0
-                keystoreForAppList=/tmp/filelist.backup_apps.list
-                olddir="$PWD"
-                #cd $keystorePath && $BUSYBOX find | grep "${USERID}_" > $keystoreForAppList
-                cd $keystorePath && sudo $BUSYBOX find -name "*${USERID}_*" > $keystoreForAppList
-                cd "$olddir" &>/dev/null
-                # check if there are any $USERID matches at all
-                if [ `$BUSYBOX stat -c %s $keystoreForAppList` -gt 0 ] ; then
-                    sudo $BUSYBOX tar -C $keystorePath -czpf - -T "$keystoreForAppList" 2>/dev/null | pv -trabi 1 > keystore_${dataDir}.tar.gz
+                if $DO_ACTION_APK ; then
+		    sudo $BUSYBOX tar -C $DATA_PATH/$appDir -czpf - ./ 2>/dev/null | pv -trabi 1 > "$appPackage"
                 fi
-                rm $keystoreForAppList
+                # get data
+                if $DO_ACTION_DATA ; then
+		    sudo $BUSYBOX tar -C $DATA_PATH/data/$dataDir -czpf - ./ 2>/dev/null | pv -trabi 1 > "$(getDataFileName "${appPackage}")"
+                fi
 
-                if $DO_BACKUP_EXTRA_DATA ; then
+                if $DO_ACTION_KEYSTORE ; then
+                    # get keystore if exists
+                    AS=""
+                    USERID="`getUserId  $DATA_PATH/data/$dataDir`"
+                    keystorePath=$DATA_PATH/misc/keystore/user_0
+                    keystoreForAppList=/tmp/filelist.backup_apps.list
+                    olddir="$PWD"
+                    #cd $keystorePath && $BUSYBOX find | grep "${USERID}_" > $keystoreForAppList
+                    cd $keystorePath && sudo $BUSYBOX find -name "*${USERID}_*" > $keystoreForAppList
+                    cd "$olddir" &>/dev/null
+                    # check if there are any $USERID matches at all
+                    if [ `$BUSYBOX stat -c %s $keystoreForAppList` -gt 0 ] ; then
+                        sudo $BUSYBOX tar -C $keystorePath -czpf - -T "$keystoreForAppList" 2>/dev/null | pv -trabi 1 > "$(getKeystoreFileName "${appPackage}")"
+                    fi
+                    rm $keystoreForAppList
+                fi
+
+                if $DO_ACTION_EXT_DATA ; then
                     extraDataPath="$DATA_PATH/media/0/Android/data/${dataDir}"
-		    sudo $BUSYBOX tar -C $extraDataPath -czpf - ./ 2>/dev/null | pv -trabi 1 > extradata_${dataDir}.tar.gz
+		    sudo $BUSYBOX tar -C $extraDataPath -czpf - ./ 2>/dev/null | pv -trabi 1 > "$(getExtraDataFileName "${appPackage}")"
+                fi
+
+                if $DO_ACTION_PERMISSIONS ; then
+                    getPermsXmlData "$dataDir" > $(getPermFileName "${appPackage}")
                 fi
